Main group 18 customer/orders

+++ back-end/src/database/migrations/20220822223441-create-sale.js
@@ -26,7 +26,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(4, 2)
       },
       delivery_address: {
         allowNull: false,
